feat(ports): Auto-add bit selection for ports. If a port has a smaller width that the external wire it connects to, then auto-add bit selection of lowest bits in external wire when connecting to the port.
Currently not working well for ports that use verilog parameters as widths.
For example, a submodule's port has width "ADDR_W", then the port
connection in the top will also be "[ADDR_W-1:0]", however, the top may
not have the "ADDR_W" parameter (or may have one with different value).