feat(axi_merge): Fix bug in busy detection. Fix bug that cause the merge to switch masters when a new one asserts valid while the previous one had already asseted valid but not yet received ready. With this commit, when a master asserts valid, the merge becomes busy without waiting for ready.
Note that we cannot switch masters after the address valid has been asserted,
because the slave would first receive one address and then receive
another one without having sent the first one's ready. And that is
against the AXI specification.

Here is a quote from the AXI specification for the Snoop channel (uses
same Handshake rules as the other channels):
"""
The standard AXI VALID/READY handshake rules apply. See Handshake process on page A3-37.
When the ACVALID signal is asserted the snoop address and control signals on ACADDR, ACPROT, and
ACSNOOP must not change until ACREADY is asserted by the master. When ACVALID is asserted, it must
remain asserted until ACREADY is asserted.
"""

Co-authored-by: Andre Nobrega <29956896+And-Nob@users.noreply.github.com>