feat(iob_system): Add support for CPU = "none". If the "cpu" python parameter of iob_system is set to "none", the system will remove memories and crossbar components. It will add the `iob_s` port that connects to the system's peripheral bus. This allows the port to be connected to an external CPU/testbench, giving direct access to system's peripherals.
Also added the `iob_system_tester_no_cpu` as a new superblock of
iob_system to use as an example.
This new tester is generated in the `submodules/tester_no_cpu/` folder
of the build directory.
The `iob_system_sim.py` (simulation wrapper) was also updated to export
this new `iob_s` port to the testbenches. (It is not driven currently).

Related to issues:
- https://github.com/IObundle/py2hwsw/issues/115
- https://github.com/IObundle/py2hwsw/issues/74