feat(iob_csrs): Remove 'VERSION' macro Resolves https://github.com/IObundle/py2hwsw/issues/153
This commit removes the 'VERSION' macro from the iob_csrs, to prevent it
from appearing in the generated `iob_uart_csrs_conf.vh` file.

The `iob_uart_version.vh` file is still generated as this one is created
by py2hwsw for all cores.

The version register is also being generated correctly in the `*_csrs.v`
files.
For example, in the `ìob_uart_csrs.v`  file, the version register is
implemented as:
`rdata_nxt[16+:16] = 16'h0081 | 16'd0;`