fix(axis2ahb): Fix bugs Fix AHB memory widths; Fix axis2ahb state values; Remove axis2ahb hwdata pipe. This delay does not seem to be needed and was breaking the write transactions;
TODO: Fix axis2ahb read transactions.
It seems that the axis2ahb has a bug in the read transactions, where it
asserts a new 'address' and 'trans' in the same clock as it tries to
read the rdata value for that address. The rdata for an address will
only come, at least, one clock cycle later than the address assertion.

I tried adding a new state between the 'WAIT_DATA' and 'TRANSFER'
states, but it did not work.

Currently the test is failing because it is reading the previous value
from the one intended (since it is reading rdata from one clock sooner than it
should).