feat(tb): Use named pipe to communicate from V to C, to improve efficiency.
Similar to the case of the console.py, by using a named pipe, the `tb`
process no longer needs to keep polling the v2c.txt file. This decreases
the amount of CPU time consumed by the `tb` process.

Also fixed a few bugs in the `tb` code, like the "fscan_ret" if statements incorrectly
checking for 4 bytes instead of 5.