fix(nix): Add missing jre dependency for libreoffice of nix environment Should resolve https://github.com/IObundle/py2hwsw/issues/213
I was not able to reproduce the error:
`javaldx: Could not find a Java Runtime Environment!`
Likely because all machines I've used have Java installed.
However, it seems that the nix 'libreoffice' package does not bring java
as a dependency. This commit adds this missing dependency to the nix environment.