feat(ipxact): Add `--browse` flag to py2hwsw: Generates ip-xact of lib cores Related to https://github.com/IObundle/py2hwsw/issues/85
Calling `py2hwsw` with the `--browse` flag will generate the
"ipxact_lib" folder in the current working directory with the IP-XACT files for all py2hwsw lib cores.

Also added 'lib-ipxact' Makefile target to generate this folder.

WIP: When trying to include this library in kactus2, it crashes with a
segmentation fault.