fix(ipxact): Generate IP-XACT with field properties from iob_csrs. Resolves https://github.com/IObundle/py2hwsw/issues/234 Resolves https://github.com/IObundle/py2hwsw/issues/242
The `csr_classes.py` script contails the `csr_field` class with
properties for each field.
There are a few checks to ensure the fields do not overlap within the
same CSR; make sure field volatility and type does not contradict CSR volatility and type, etc.

If a CSR is created with no fields, a single field is automatically created with the
same properties as the CSR.

CSRs are assumed to have widths multiple of 8 bits. If fields are not
specified, for all bits, then 'RSVD' (reserved) fields are automatically
created for them.