wip(api): Add idea of using @api_for decorator in every API class.
With this decorator, two objects should be created: API object and internal object.

The user only has access to the API object. It only contains attributes and methods defined in the API class.
The internal object is only used internally by py2. It includes the attributes and methods of the API object, but also extra functionality specific to the internal class.

Py2 should convert from one type of object to the other every time it exchanges objects with the user.