wip(api): Fix API methods.
Add example in iob_and.py, where we try to call the API's 'test_method()' of iob_conf.
This will throw an error message, as expected, since iob_conf does not
have the implementation of this method.
Adding a `test_method()` to iob_conf fixes this, and the method runs as expected.