Fix python env in default.nix
Not sure why, but generating a python environment with the nix `pkgs.python3.withPackages` seems to work better than placing the python packages in the buildInputs.

This fixes issue where python packages were missing from the environment, even though they were in the mkShell buildInputs.

Also, doing `echo $PYTHONPATH` is not showing the packages from this environment, but they are successfully imported in python.

Line 142 of builtins.trace no longer shows python package versions with this commit. We should fix this line to keep version up to date.