fix(shell): Workaround for EOFError in non-interactive mode
Normally, the last line of the file is not evaluated if EOF is received.
This workaround sends one last blank line to stdin to force the last
line to be evaluated, when running non-interactively.

For example, if the file ends with:

if __name__ == "__main__":
	some_code

The code inside that if statement is not evaluated, because the if block
is not terminated, due to the EOF.