wip(port): Make 'direction' be a virtual property of iob_port
We've decided that the direction suffix of a wire/bus may stay stored in
its 'name' attribute.
This allows us to distiguish between two wires named 'data_i' and
'data_o' in the same core.

Since the name of the wire/bus already stores this information, we dont
need a dedicated 'direction' attribute in the iob_port class anymore.
Plus, having this attribute would be duplicating direction info, which can lead to
bugs.

This commit introduces a setter and getter for the virtual 'direction'
property of iob_port. It obtains the direction from the name suffix of the
associated wire/bus. And changes direction by changing that name suffix.

Using python code, if we create a port associated to a previsouly
created wire, we will either have to add the direction suffix to that
wire, or use the port's direction setter to do it for us.