fix(tb): Fix main loop in iob_v_tb.v
Resolves https://github.com/IObundle/py2hwsw/issues/369

Main loop of iob_v_tb.v was advancing simulation, even when no requests
were made (from iob_core_tb.c). This would allow simulation to advance
in slow machines, skipping important control checkpoints.

This commit removes the time advance in the main loop when no requests
are received.
Note: This means that the `iob_core_tb.c` must keep making requests in
order for the simulation to advance. I'm not sure if this can be a
problem for some testbenches.