feat(tb): Use named pipe for c2v.txt (more efficient)
By using named pipe for the c2v.txt communication, the testbench no
longer needs to poll the file. It can just make a blocking request to
read data from the pipe. This is more efficient than wasting CPU time by
polling the file.

This implementation is an extension of solution proposed in
PR https://github.com/IObundle/py2hwsw/pull/370.
It has the same caveats (the iob_core_tb.c must make requests to the
testbench for the simulation to advance).