fix(format): Dont format build dir when formatting current repo
Py2 has 2 separate formatting steps:
- One for the current repo (for example py2hwsw if running the lib Makefile).
- One for the build directory.

If the build directory is inside the current repo, we dont want to
format it twice.
We specifically dont want to format the build directory when formatting
the current repo, since the build dir may have some specific files that
must not be modified (Like the `libfdt.h` of iob_linux).