fix(csrs): Maintain address stable until the next valid
This allows current register to maintain its rdata output, even after rvalid
until the next valid comes.
It seems that the lack of iob_rready does make a difference because if
we dont maintain the rdata output for more than one cycle (even after
rvalid), the system does not work. Maybe there is a bug somewhere else
in the system? Maybe in axi2iob converter?