fix(lib): iob_split and iob_merge modules. Update lib modules.
Iob_split and iob_merge selection signals had an issue when a new valid
came before the previous rvalid was sent. This could cause the
split/merge to switch manager/subordinate interfaces before the
transaction completed.

To fix this, a FSM was added to track the current state of the
transaction, blocking new handshake signals until the current transaction
is complete.

- Rename 'output' and 'input' interfaces in iob_split and iob_merge to refer to
  standard 'manager' and 'subordinate' interfaces.

- Update other lib modules that include iob_split and iob_merge to use new
port names and python paramters.